Add files via upload
In this project, the goal was to analyze the sales data of a store and based on that analysis, firstly predict the expected revenue for the next two years considering the current sales situation of the store, and secondly, show how a slight change in some parameters, like converting a small percentage of Promiscuous customers to Loyal ones, could potentially increase future revenue.
The project was completed in three main parts:
Data Wrangling: In this phase, the dataset was examined and corrected for issues like duplicate data, incorrect data types, missing values, and standardization of data such as aggregating products that were defined in two or more departments into their main department. Eventually, a cleaned dataset was produced.
Data Analysis: Using the cleaned dataset, we analyzed the current sales situation. This included analysis of products, departments or product categories, months with the highest and lowest sales, customer purchase and profitability analysis based on their loyalty, overall sales trends of the store, and a heatmap of correlation between different sections of the dataset, like Product Preference based on Loyalty.
Forecasting: This part guides through forecasting total revenue with time series analysis. It covers data loading, preprocessing, and monthly aggregation. Key steps include exploratory analysis with seasonal decomposition, model selection (ARIMA or SARIMAX), and fitting the model to the training data. It emphasizes model evaluation using statistical metrics, diagnostic checks, and forecasting future revenue.